fix: remove market price bug on currency change
