+++ src/pages/Swap/LimitOrderBox/components/OrderLimitPriceField/OrderLimitPriceField.tsx
@@ -1,13 +1,14 @@
 import { formatUnits, parseUnits } from '@ethersproject/units'
 import { Token, TokenAmount } from '@swapr/sdk'
 
-import { useContext, useEffect, useState } from 'react'
+import { useCallback, useContext, useEffect, useState } from 'react'
 import { RefreshCw } from 'react-feather'
 import { useTranslation } from 'react-i18next'
 
 import { invalidChars } from '../../constants'
 import { LimitOrderFormContext } from '../../contexts/LimitOrderFormContext'
-import { LimitOrderKind, MarketPrices } from '../../interfaces'
+import { LimitOrderKind } from '../../interfaces'
+import { calculateMarketPriceDiffPercentage } from '../../utils'
 import { InputGroup } from '../InputGroup'
 import {
   LimitLabel,
@@ -43,7 +44,7 @@ export function OrderLimitPriceField({ id }: OrderLimitPriceFieldProps) {
   const toggleCurrencyButtonLabel = `${quoteTokenAmount?.currency?.symbol}`
   const [inputLimitPrice, setInputLimitPrice] = useState<string | number>(formattedLimitPrice)
 
-  let { marketPriceDiffPercentage, isDiffPositive } = calculateMarketPriceDiffPercentage(
+  const { marketPriceDiffPercentage, isDiffPositive } = calculateMarketPriceDiffPercentage(
     limitOrder.kind,
     marketPrices,
     formattedLimitPrice
@@ -59,7 +60,7 @@ export function OrderLimitPriceField({ id }: OrderLimitPriceFieldProps) {
   /**
    * Handle the base currency change.
    */
-  const toggleBaseCurrency = () => {
+  const toggleBaseCurrency = useCallback(() => {
     // Toggle between buy and sell currency
     const kind = limitOrder.kind === LimitOrderKind.SELL ? LimitOrderKind.BUY : LimitOrderKind.SELL
     const [baseTokenAmount, quoteTokenAmount] =
@@ -76,14 +77,14 @@ export function OrderLimitPriceField({ id }: OrderLimitPriceFieldProps) {
     const nextLimitPriceFormatted = nextLimitPriceFloat.toFixed(6) // 6 is the lowest precision we support due to tokens like USDC
     const nextLimitPriceWei = parseUnits(nextLimitPriceFormatted, quoteTokenAmount?.currency?.decimals).toString()
 
-    setLimitOrder({
-      ...limitOrder,
+    setLimitOrder(oldLimitOrder => ({
+      ...oldLimitOrder,
       kind,
       limitPrice: nextLimitPriceWei,
-    })
+    }))
     // update the formatted limit price
     setFormattedLimitPrice(nextLimitPriceFormatted)
-  }
+  }, [buyTokenAmount, limitOrder, sellTokenAmount, setFormattedLimitPrice, setLimitOrder])
 
   /**
    * Handle the limit price input change. Compute the buy amount and update the state.
@@ -132,14 +133,19 @@ export function OrderLimitPriceField({ id }: OrderLimitPriceFieldProps) {
       setBuyTokenAmount(nextTokenBuyAmount)
       setFormattedLimitPrice(nextLimitPriceFormatted)
       setFormattedBuyAmount(newBuyAmountAsFloat.toString())
-      setLimitOrder({
-        ...limitOrder,
+      setLimitOrder(oldLimitOrder => ({
+        ...oldLimitOrder,
         limitPrice: parseUnits(nextLimitPriceFormatted, quoteTokenAmount?.currency?.decimals).toString(),
         buyAmount: nextBuyAmountWei,
-      })
+      }))
     }
   }
-
+  console.log('toggle out', {
+    limitOrderKind: limitOrder.kind,
+    formattedLimitPrice,
+    baseTokenAmount,
+    quoteTokenAmount,
+  })
   return (
     <InputGroup>
       <LimitLabel htmlFor={id}>
@@ -181,34 +187,4 @@ export function OrderLimitPriceField({ id }: OrderLimitPriceFieldProps) {
       </InputGroup.InnerWrapper>
     </InputGroup>
   )
-}
-
-export function calculateMarketPriceDiffPercentage(
-  limitOrderKind: LimitOrderKind,
-  marketPrices: MarketPrices,
-  formattedLimitPrice: string
-) {
-  const nextLimitPriceFloat = limitOrderKind === LimitOrderKind.SELL ? marketPrices.buy : marketPrices.sell
-  let marketPriceDiffPercentage = 0
-  let isDiffPositive = false
-
-  if (Boolean(Number(nextLimitPriceFloat))) {
-    if (limitOrderKind === LimitOrderKind.SELL) {
-      marketPriceDiffPercentage = (Number(formattedLimitPrice) / Number(nextLimitPriceFloat.toFixed(6)) - 1) * 100
-      isDiffPositive = Math.sign(Number(marketPriceDiffPercentage)) > 0
-    } else {
-      marketPriceDiffPercentage = (Number(nextLimitPriceFloat.toFixed(6)) / Number(formattedLimitPrice) - 1) * 100
-
-      if (marketPriceDiffPercentage < 0) {
-        marketPriceDiffPercentage = Math.abs(marketPriceDiffPercentage)
-      } else {
-        marketPriceDiffPercentage = Math.min(marketPriceDiffPercentage, 999)
-        marketPriceDiffPercentage = marketPriceDiffPercentage * -1
-      }
-      isDiffPositive = Math.sign(Number(marketPriceDiffPercentage)) < 0
-    }
-  }
-
-  marketPriceDiffPercentage = Math.min(marketPriceDiffPercentage, 999)
-  return { marketPriceDiffPercentage, isDiffPositive }
 }
